Create setDesktop.zsh
This script will do an initial set of the user's desktop. It makes use of the utility
desktoppr created by Armin Briegel at scriptingosx (https://github.com/scriptingosx/desktoppr).

After this script runs it will not run again for non administrtor uses because it checks 
for a desktoppr file placed in the following location: $HOME/Library/dockutil.
This way users can further modify the dock to their desires.

This needs to run in conjunction with a LaunchAgent to get it to run at login.